Added conditonal to enable the certificate filter
