+++ src/main/resources/application.properties
@@ -9,6 +9,8 @@ spring.datasource.password=postgres
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
 
+certificate.filter.enabled=false
+
 aws.access.key=${AWS_ACCESS_KEY_ID}
 aws.secret.key=${AWS_SECRET_ACCESS_KEY}
 aws.region=eu-west-1
